moved parse_checksum method to utils.models
