+++ src/piddiplatsch/utils/models.py
@@ -3,6 +3,7 @@ from datetime import datetime
 from uuid import NAMESPACE_URL, uuid3
 
 from dateutil.parser import isoparse
+from multiformats import multihash
 
 from piddiplatsch.config import config
 
@@ -58,3 +59,31 @@ def parse_pid(value: str) -> str:
     else:
         pid_ = value
     return pid_
+
+
+def parse_multihash_hex(checksum_with_type: str) -> tuple[str, str]:
+    """
+    Parse a multihash hex string into (checksum_type, checksum_hex).
+
+    Args:
+        checksum_with_type: multihash as hex string, e.g.
+            '12205994471abb01112afcc18159f6cc74b4f511b99806da59b3caf5a9c173cacfc5'
+
+    Returns:
+        Tuple of:
+        - checksum_type: human-readable hash name, e.g., 'sha2-256'
+        - checksum_hex: raw digest as hex string
+    """
+    # Convert hex string to bytes
+    mh_bytes = bytes.fromhex(checksum_with_type)
+
+    # Decode the multihash
+    code, digest_bytes = multihash.unwrap_raw(mh_bytes)
+
+    # Get human-readable hash name
+    checksum_type = multihash.get(code).name
+
+    # Convert digest bytes to hex
+    checksum_hex = digest_bytes.hex()
+
+    return checksum_type, checksum_hex
